lesson1

+++ src/lesson03/hometasks/Task4Weather.java
@@ -2,8 +2,9 @@ package lesson03.hometasks;
 
 public class Task4Weather {
 
+    // реализация через метод с возвратом значений
     public double findAverageTemp(int monday, int tuesday, int wednesday, int thursday, int friday, int saturday, int sunday){
-        double averageTemperature = (int)(monday + tuesday + wednesday + thursday + friday + saturday + sunday) / 7;
+        double averageTemperature = (double) (monday + tuesday + wednesday + thursday + friday + saturday + sunday) / 7;
          return averageTemperature;
 
 }
